refactor(e2e_harness): extract federated SQL assembly into query_build.go
Pure move of the builders this PR touched (dynamic entry points, combined
query, parquet tier + row_id-semijoin pushdown, post-dedup final
select/count, dedup CTE) — the first slice of the #220 query.go split
along its planned tier-query-construction seam. query.go drops from 1237
to 1067 lines (net -121 vs main); the remainder of the split stays #220.
